internal: Pass job ID during artifact completion. (#604)
* Merge subscribers.

* Pass job id.

* Trigger CI.

* Add keywords.
