+++ crates/core/action-pipeline/src/pipeline.rs
@@ -2,8 +2,7 @@ use crate::errors::PipelineError;
 use crate::processor::process_action;
 use crate::run_report::RunReport;
 use crate::subscribers::local_cache::LocalCacheSubscriber;
-use crate::subscribers::moonbase_cache::MoonbaseCacheSubscriber;
-use crate::subscribers::moonbase_ci::MoonbaseCiSubscriber;
+use crate::subscribers::moonbase::MoonbaseSubscriber;
 // use crate::worker_pool::WorkerPool;
 use console::Term;
 use moon_action::{Action, ActionStatus};
@@ -384,22 +383,10 @@ async fn create_emitter(workspace: Arc<RwLock<Workspace>>) -> Emitter {
             }
         }
 
-        if let Some(session) = &local_workspace.session {
-            if session.remote_caching_enabled {
-                emitter
-                    .subscribers
-                    .push(Arc::new(RwLock::new(MoonbaseCacheSubscriber::new())));
-            } else {
-                MoonbaseCacheSubscriber::not_enabled();
-            }
-
-            if session.ci_insights_enabled {
-                emitter
-                    .subscribers
-                    .push(Arc::new(RwLock::new(MoonbaseCiSubscriber::new())));
-            } else {
-                MoonbaseCiSubscriber::not_enabled();
-            }
+        if local_workspace.session.is_some() {
+            emitter
+                .subscribers
+                .push(Arc::new(RwLock::new(MoonbaseSubscriber::new())));
         }
     }
