+++ crates/core/action-pipeline/src/subscribers/mod.rs
@@ -1,3 +1,2 @@
 pub mod local_cache;
-pub mod moonbase_cache;
-pub mod moonbase_ci;
+pub mod moonbase;
